updates README and cli tools to include update endpoint and to put together the CLI as a class with subcommands
